project-main/requirements.txt
# Django Core
Django==4.2.0
djangorestframework==3.14.0

# Local development database (SQLite)
# No need for psycopg2-binary for local

# Security & Auth
django-guardian==2.4.0
django-cors-headers==4.0.0
Pillow==9.5.0

# File handling
python-magic==0.4.27

# Environment
python-decouple==3.8

# Development tools
django-extensions==3.2.1